% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate_matrix.R
\name{triangulate_matrix_na}
\alias{triangulate_matrix_na}
\title{Make Reduced Matrix Size}
\usage{
triangulate_matrix_na(heightmap, maxError = 0, maxTriangles = 0)
}
\arguments{
\item{heightmap}{A two-dimensional matrix, where each entry in the matrix is the elevation at that point.
All points are assumed to be evenly spaced.}

\item{maxError}{Default `0`.}

\item{maxTriangles}{Default `0`, which turns off this setting.}
}
\description{
Uses Delaney triangulation to reduce the size of the matrix
}
\keyword{internal}
